+++ mobile/parent_app/lib/core/router/routes.dart
@@ -7,5 +7,6 @@ class Routes {
   static const login = '/login';
   static const home = '/';
   static const profile = '/profile';
+  static const profileEdit = '/profile/edit';
   static String childDetail(int id) => '/children/$id';
 }